Downgrade rdkafka back to 0.23.1
rdkafka 0.26.0 depends on tokio 1.0, which means it breaks if we ever
attempt to actually use it due to not being run in the context of a
tokio 1.0 refactor. We should really upgrade our tokio version (and also
maybe write an integration test to cover this!) but in the meantime we
just have to downgrade for now.

This also removes the cmake_build feature, which was necessary to get it
building on my machine - I'm not sure why it was there in the first
place, since there's no real commit history for the debezium_connector
pre subtree merge, but this seems to work fine (I've even deployed this
version to AWS successfully).