Only pass supported DDL through the dbz connector
Debezium sends quite a bit of DDL on the schema change topic that we
don't know how to handle - namely some DROP TABLE IF EXISTs at the
beginning, some SETs, etc. Unfortunately, it's not trivial right now to
robustly classify which of those statements we actually don't know how
to handle, and we can't just send them over to noria because we aren't
able to figure out if the resulting error was due to the statement
failing to parse or something more severe. For now, this
just *textually* looks at the ddl with a *regex* (this makes me very
upset) to determine if it's a CREATE TABLE or a CREATE VIEW, and ignores
everything else.

Ugh.