Implement a bunch more integer-type coercions
Implement a much wider swath of integer-type coercions for DataType,
including both bigint and int going to {unsigned,}{tiny,small}int. Per
the comment in the code, we don't check the "length" field on the type,
since mysql doesn't actually use that as a max type length - it actually
just controls display.