Move aggregates after filters
The semantics of SQL state that all aggregate operations happen
logically *after* filters - so that queries like `SELECT count(*) FROM
t1 WHERE x > 5` give eg "the number of rows in t1 where x is greater
than 5", rather than "the number of rows in t1, but then filter by x".
My best guess as to why we *used* to think putting aggregates before
filters was okay is because (I assume) we only used to support direct
equality filters - with those, we just add the column we're filtering on
to the GROUP BY of the aggregate - if it's just an equality, that
actually ends up getting the correct results!

Regardless, there are more types of filters than just equality, and any
non-equality filter will give the wrong results (specifically
under-grouped rows) if filters are placed after aggregates, so this
moves aggregates above filters.

That revealed an incorrect usage of SQL in a couple of tests, that were
using WHERE to filter on the results of aggregates, which SQL doesn't
allow (you have to use HAVING) - since that was broken before (albeit in
a different way which the user could work around) and we don't
actually *support* HAVINg yet (we parse it, but ignore it) I've marked
those tests as ignored for now (accepting the slight regression) and
opened a clubhouse ticket (in a comment in the code) to implement HAVING