Zookeeper retry mechanism.
Added an exponential backoff retry mechanism for Zookeeper.
There are two things to be said about this:
1. Currently, the Zookeeper::connect, the Zookeeper library method we call when
creating a ZookeeperAuthority, won't fail if Zookeeper is not available. To
workaround that, we make a call to Zookeeper::exists just to try to reach
Zookeeper. The downside of this, is that that method call might take longer than
 the duration specified in Zookeeper::connect. For more information, see
 https://github.com/bonifaido/rust-zookeeper/issues/64.
2. We had two choices of libraries that already provide retry mechanisms:
`retry` and `backoff`. Although `backoff` has a version `0.3.0` (versus `retry`
which has a version `1.2.0` and seems more stable), by looking at the code,
`backoff` has a better exponential backoff algorithm than `retry`, which is
extremely naive.