ast: Replace FunctionArgument with Expression
Replace the FunctionArgument type, which is an ad-hoc expression AST
node that contains a strict subset of the AST nodes present in an
arbitrary SQL expression, with the new Expression enum. Also, move
FunctionExpression to expression.rs (it never really belonged in
column.rs) and get rid of the FunctionArguments type in favor of just
using a Vec<Expression>.

This is the first step in many of replacing all of the ad-hoc, one-off
expression AST enums in the nom-sql AST with Expression, to reduce
nesting, simplify pattern matching on and constructing the AST, and
reduce the amount of duplication in code that has to traverse the
AST (and pave the way for creating a Visitor!). Since this is the first
time we're actually using Expression inside of nom-sql itself, this also
slightly tweaks the representation of Expression to make it fit better
in how we need to use it (and also make the refactor more gradual, by
using Column directly instead of inlining a subset of the fields in a
struct enum variant).

On the MIR side, since function calls as arguments to other function
calls are no longer being referenced as `Column`s (they're just
Expression::Call nodes now) we have to explicitly track the names that
are generated for expressions that we project above aggregate columns,
and map the expressions themselves to those names. This is a little
messy right now, but actually paves the way very well to the refactor to
the *MIR* Column type that's happening in parallel to this work.

Finally, in the interest of making the refactor more gradual, I've
elected *not* to actually change the parser itself at this stage,
instead just updating the types for the existing function_argument
parser - at some point we'll want to have an `expression` parser, that
we use in as many places as possible, but that can come later.