Clarify data type field in noria-psql Resultset and Row, fix latent lookup bug.
Rename col_types to project_field_types in Resultset and Row, clarifying that
the sequence of fields referenced parallels the sequence in project_fields and
not the sequence in values (which may contain additional non-projected fields).

Correct the field type index lookup in RowIterator's next(), which had been
performing a lookup using the values index rather than the project_fields index.
(It appears that the previous behavior could not actually cause an error because
non-projected fields always appear at the end of a row. Nonetheless it has been
fixed.)

Additionally, add documentation and unit tests.