ast: Simplify CASE WHEN expressions
Inline the CaseWhenExpression type as a struct enum member into the
Expression enum, and replace the one-off ColumnOrLiteral type it was
using for the `then` and `else` branches with Expression.

This was also a lovely opportunity to bite the bullet and add an
`expression` parser, which is now used both for the body of a CASE
WHEN and for function arguments (`function_argument_parser` is gone!).