Make various aggregates into structs
Various aggregate functions (Avg, Count, Sum, GroupConcat) that either
can be DISTINCT or (in the case of GroupConcat) take an argument were
previously positional tuple structs, which is confusing especially in
the case of the DISTINCTs because the second argument was just an opaque
boolean. This refactors those to all be struct enums instead, which
greatly improves clarity.