upgrade mysql_async dependency for noria client
mysql_async dependency, which is a fork owned by readyset, upgraded to
be based on mysql_async v0.27.1, previously v0.23.1. This is neccessary
for compatibility with tokio 1.x runtimes.

Additionally, some of the client can now be refactored to take advantage
of other changes introduced in newer mysql_async, such as not consuming
the Conn during queries. But this work is left for a future PR