ast: Pull aliases up to FieldDefinitionExpression
Rather than storing aliases for fields in a select query all over the
place  - including in some places where they're not actually
syntactically valid! - make `alias` a field of a new variant of
FieldDefinitionExpression called "Expression", which subsumes both Col
and Value, and has an Expression field and an Option<String> alias
field. This brings FieldDefinitionExpressions basically to the place
where I think they should be, modulo the name which I think is a little
cumbersome.

A lot of stuff had to change to make this happen - most importantly,
since aliases are no longer getting stored on nom_sql::Column we have to
be careful to carry them through the query_graph to MIR - to make this
happen, I've added an `alias` field to OutputColumn::Data, which
previously just had a column.

Conveniently (though perhaps slightly increasing the scope of this
change), since the primary data type on a FieldDefinitionExpression
is now actually an Expression, we are no longer parsing top-level
function calls as nom_sql::Columns with a `function` attribute, but as
instances of the Expression::Call variant. This gets us almost all the
way to removing the `function` attribute from nom_sql::Column entirely,
which is awesome, except that there's still a *couple* of places where
we still have Columns with a function attribute - notably, in
ArithmeticBase. Because of the extra complexity involved in
conditionally dispatching between Expression::Call nodes and function
columns in that instance, I've elected to (temporarily) disable
the *one* test that depends on that functionality for now, and once we
have the opportunity to replace ArithmeticBase with Expression entirely
we can re-enable it.