Get rid of ArithmeticExpression
Now that we're storing aliases on FieldDefinitionExpression::Expression,
there's no need to have an extra ArithmeticExpression type wrapping an
Arithmetic and only adding an alias - not to mention that the only place
that was using it (ConditionExpression::Arithmetic) didn't really make
sense to have an alias anyway! So we can make that just use an
Arithmetic and get rid of the data type entirely.