ast: Replace Arithmetic{Item,Base} with Expression
In the continuing saga of unifying and simplifying the nom-sql AST,
replace both ArithmeticItem and ArithmeticBase (used in various
different highly confusing methods as the lhs/rhs of an Arithmetic) with
just Expression. To simplify the parser while also integrating it with
the rest of the expression parser overall, this also redefines the token
tree + operator precedence parsing for arithmetic expressions (not yet
conditional expressions) to use a Pratt parser, using the (imo very nice
and easy to use) `pratt` rust crate.

The parsing for ConditionalExpression's arithmetic branch unfortunately
had to add a bit of a hack here as a result, since it's an entirely
different Expression type so it needs to *embed* arithmetic parsing as
an entirely separate parser, meaning we need to first parse an
arithmetic expression, then pattern match on it to make sure it is in
fact an arithmetic expression. This is quite unfortunate, but is only a
temporary measure until we can merge ConditionExpression and Expression
entirely.