Bring back expressions over aggregates
Now that we've refactored arithmetic to use Expressions on the left and
right-hand side, we can much more simply implement the part of the
query_graph that scans expressions for aggregates to only actually refer
to expressions - here, we approach that by recursively walking an
expression turning all function call nodes for aggregate functions
into *columns* referencing those aggregate functions, and spitting out a
vector giving the functions themselves and the names they map to. This
allows us to de-#[ignore] the integration test for arithmetic on
aggregates, and also (un-tested, for now) implements *other* sorts of
expressions on aggregates, like function calls!