Omit empty groups for COUNT + GROUP_CONCAT
If a deletion results in no rows for a particular group in a COUNT or a
GROUP_CONCAT, only output the negative for the group, omitting the
positive for the row with a 0, matching the behavior of MySQL when
there are no rows in an aggregate. This is implemented by adding a new
`empty_value()` method to the GroupedOperation trait, which can
optionally specify the empty value (the identity of the monoid provided
by aggregates) for the aggregate. Since averages maintain some extra
internal state for the count of the rows, this doesn't work there
unfortunately - we'll have to find a different strategy to fix that.