Implement support for timestamps in js-client
Implement conversion to/from js `Date` objects and `DataType::Timestamp`
types (which are backed by `NaiveDateTime` timestamps).

Note: there was a decision made here of how to convert a `NaiveDateTime`
timestamp into a js `Date` object due to the ambiguity of
`NaiveDateTime`'s lack of a timezone. We decided to interpret all
timestamps as UTC timestamps. This is what the postgres node.js client
does. The default behavior for the mysql node.js client conforms to the
mysql server's local timezone, but we don't intend to make that
configurable and that should *ideally* always be set to UTC anyway.