Introduce building views from replicas based on filters.
This refactors the ViewRequest to take a `filter` parameter
that can filter for a specific view. ViewBuilder's can then take
the ViewRequest and apply filters when creating a view over
the set of reader replicas for a query.

A new filter parameter is added `Region` which is used for selecting
readers in a specified region if possible. The `Region` filter will
be implemented in a future change and is currently a no-op.