Worker support for specifying their region for reader replica selection
This change adds a command line argument to noria-server, `region`,
that allows a worker to specify the region that they live in. This
region is propagated to the controller to allow client requests to
route requests based on noria-server's regions.

This commit does not change any request routing and still selects
the first reader replica found when performing a breadth first search
of the subtree of a specific query.