Implement new ColumnSource API, removing parent_columns / resolve
As discussed in the design doc for fully partial joins, this commit
implements the new `Ingredient::column_source` API, and removes
`parent_columns` and `resolve` from ingredient implementations.

These methods are then reimplemented in terms of `column_source`, so
the rest of the code continues to work as it did before.

The old design was always quite hacky:

- the difference between `resolve` and `parent_columns` wasn't really
  specified very well, making ingredient authors confused as to how
  the two should differ, if at all
- the information returned by these functions was used to generate
  replay paths in `provenance_of` (in `keys.rs`), but how replay paths
  were constructed from said information was unclear, and required
  additional hacks / special-cases (e.g. `on_join`) to accommodate
  ingredient-specific behaviour
- the old design didn't provide any room for generated columns, which we
  will shortly be implementing

This is now done away with, with the `ColumnSource` enum letting
ingredients provide complete information to the materialization planner;
the `on_join` logic, for example, is now actually in the `Join`
ingredient implementation. Yay separation of concerns!

The semantics of `must_replay_among` have also changed slightly to
reflect the fact that `on_join` is no more (ingredients must now choose
what replay paths they want, instead of letting the materialization
planner choose among a set of options).

We include the `handle_upquery` API from the design doc, even though
it's currently dead code; that will be implemented in a future commit.