Made GroupConcat part of Aggregation(in MIR land)
In order to simplify MIR, it makes sense for GroupConcat to be a part of
aggregation rather than its own Node Type. This will be useful in a
future commit when implementing a columns() function.

In addition, I had to fix the casing on the Aggregation enum to make
cargo fmt happy.