Overhaul noria-server's coordination and RPC system
This commit straightens out the spaghetti code responsible for starting
up a `noria-server` instance, as well as the mechanisms used by the
various sub-components of a `noria-server` to communicate with one
another.

Starting a server previously involved a tangled mess of various async
functions calling one another and forking off Tokio tasks to run in,
which was hard to follow (and also made it difficult to debug failures
in one of the functions). Furthermore, the types used for
message-passing / RPC between discrete noria-server components were
highly muddled, with a single large enum (`Event`) and a 'forwarding
loop' that received these `Event`s from a variety of sources (other
components inside the server, external clients, ...) and dispatched them
to the "worker-side" and "controller-side" of the server.

This made it very difficult to trace the flow of communication between
server components, and also propagate errors across RPC boundaries
correctly.

The new code straightens all of this out by tidying up the various async
functions into three main `struct`s: `ControllerOuter` (the "controller
side" of a server instance, handling leader elections and leadership
tasks if elected), `Worker` (the "worker side" of a server instance,
handling the running of domains), and `NoriaServer` (an HTTP server that
dispatches messages to the other two).

Instead of the interactions between these three being muddled in a large
`Event` enum, there are now separate message types for each type of
interaction:

- workers have the `WorkerRequestKind` enum
- controllers have the `HandleRequest` and `ControllerRequest` enums

While the old code would use 4+ discrete TCP servers to communicate
between the controller and worker, the new code centralizes all of this
into the `NoriaServer` HTTP server, which receives HTTP requests,
deserializes them into one of the above enums, and sends the enum to the
correct component of the server. Crucially, this mechanism is
bidirectional; the result of the request is sent back to the HTTP
server, which can in turn relay it to the requesting client; this allows
for errors to be propagated across the RPC boundary in a way that was
not possible before!

This therefore makes the code in the controller that deals with
migrations *significantly* more reliable; instead of sending messages
over TCP to workers (or, even worse, raw packets to the domains) and
hoping the task completes (or, even worse, using the incredibly cursed
`DomainReplies` system to wait for acks with absolutely zero correlation
between request and response, leading to fun race conditions), requests
from the controller to the workers / domains during migrations are now
blocking HTTP calls with errors propagated.

NOTE: The "data plane" -- interactions between domains, and
domain<->client interactions -- is importantly left untouched by this
work, given that (a) it works fine and is understandable, and (b) as it
is very much on the hot path for reads/writes, changes to it could have
adverse performance effects.

Other assorted changes that are included in this commit for one reason
or another:

- the controller<->client RPC now uses `bincode` instead of JSON, given
  that the rest of Noria uses bincode and having 2 different
  serialization formats in use is undesirable
- the "done" future returned by `start_instance` was rolled into the
  returned `Handle`; equivalent behaviour can be obtained by using
  the `Handle::wait_done` method
- some integration tests had extra `sleep()` calls inserted in order
  to reduce flakiness