Js-client end-to-end testing
End-to-end testing for the js client that runs in buildkite and can
also be run locally.

Tests are run via the docker-compose file in js-client/tests/build which
wires zookeeper, the noria server, and the js client together with each
running in its own container.

Changes to the way the noria server handles IP addresses were motivated
by the desire for hostnames to resolve to IPs to work nicely with
container networking.

Note: also switched over to yarn from npm because it was working better
in the node container and it's more performant anyway in some cases.