Add shell.nix and envrc for noria+noria-ui
Add a shell.nix and .envrc file for installing all the development
dependencies for the repo (and noria-ui, which lives in its own special
world) using nix[0] and automatically with lorri[1].

I've had this in my local dev env for a while, nobody's asked for it,
but I figured I might as well commit it just in case someone else would
benefit from it! Since it's not the Blessed Way to get dev env setup
atm (yet) it probably won't be guaranteed to be kept up-to-date as we
add new dependencies, and I'm ok with having to maintain it myself given
that.