Use replay_paths_for in the materialization planner
This updates the materialization planner to use the newly introduced
replay_paths_for API, instead of provenance_of. Mostly, this just makes
the code a bit simpler (e.g. not having to cut off the paths in plan.rs,
since the new API supports that natively, not having to unwrap columns,
etc.).

Support is added for split replay paths as part of this change, which is
slightly finicky (since split replay paths require materializing the
node with generated columns).