Fix params coercion bug for executing prepared updates & inserts
Fix a bug where params when executing prepared updates and inserts
for writes to noria directly were not being coerced to the correct
sql value before being used.

Also implemented extra checking in the js-client test suite to ensure
updates are actually going through by doing a select to explicitly check
for the updates in the updated records.