Re-enable flaky integration test for replica routing by region.
This test is being reenabled after removing the line that
checked that a reader that has not been replicated has been
placed in "r1". In reality the initial reader node can be
placed in either "r1" and "r2" and we should ensure that the
other replica is placed in the other region. With one replica
in each region we can then verify we can request using the
region's identifier.