Allow reading row-level events from the MySQL binlog
This initial PR replaces the problematic debezium connector with
a new connector that reads from the binlog directly.

Currently the connector is able to properly read and forward schema
events, write rows, update rows and delete rows events.

The way the connector works is basically the same Debezium did,
it connects as a replica to the MySQL server, and request the binlog
dump, which is then delivered line by line, either from the start
or from a previously stored position.

To test the implementation as implemented here:

* Start Zookeeper :
  docker run -d -it --rm --name zookeeper -p 2181:2181 -p 2888:2888\
  -p 3888:3888 debezium/zookeeper:1.3

* Start MySQL:
  docker run -d -it --rm --name mysql -p 3306:3306\
  -e MYSQL_ROOT_PASSWORD=debezium -e MYSQL_USER=mysqluser\
  -e MYSQL_PASSWORD=mysqlpw debezium/example-mysql:1.3

* Start Noria:
  cargo run --release --bin noria-server -- --deployment myapp\
  --no-reuse --address 127.0.0.1 --shards 0

* Start the connector:
  cargo run --bin mysql_connector -- --db-name inventory --deployment myapp\
  --mysql-user root --mysql-password debezium --binlog-file mysql-bin.000003

Fixes #[ENG-17]
https://readysettech.atlassian.net/browse/ENG-17