Allow persisting replication offsets to rocksdb
When running in a read-replica configuration, we need to save the offset
we've reached in the replication log (eg the binlog in the case of
mysql) on every replicated write, so that when we restart (potentially
after a crash!) we know where to pick up. We also need this to
be *atomic*, so that the process of writing to persistent state and the
process of updating the replication offset both happen
together (otherwise we could potentially drop committed writes if we
crashed in between the two operations). Given this, the best place to
store it is in rocksdb, and the *only* way (without a pretty significant
rearchitecture of the way we store State) to have that happen atomically
is to do that as part of `process_records`. This is *slightly*
inconvenient, as it means we have to update the State trait and pollute
things that might only care about MemoryState (which doesn't have
replication offsets), but ended up being a fairly minimal change so I'm
ok with it.

This commit implements the persistence and reading of replication
offsets for persistent state, but nothing actually *uses* that yet -
there's a notable TODO in the process branch for Base nodes that will be
filled in in the next commit