Add a SetReplicationOffset TableOperation
Add a new member to the TableOperation enum which doesn't actually touch
any rows, but sets the *replication offset* (discussed more in the
previous commit) of the base table that the TableOperation is being
applied to. This threads through the `process` method on `Base` nodes,
which now returns a new `BaseWrite` struct, encapsulating a set of
records to be written in addition to an optional replication offset.
Finally, there's some miscellanious other changes to code that was
previously assuming all table operations referred to a row.