Add a controller RPC to load replication offsets
Add a new HTTP RPC method to the controller that asynchronously queries
all domains that might contain base tables for the maximum replication
offset that's been committed to those base tables. If any of the
reported repelication offsets are from different replication logs (which
in the case of MySQL can be caused by the binlog prefix changing) this
returns an error. This will be used at startup of the replication thread
to determine where to start reading the replication log.