Desugar CTEs into subqueries
In exactly the same way as we do for nested subqueries, desugar (simple,
non-recursive) CTEs to separate non-leaf queries, rewriting the *name*
of the CTE, both in the graph and (piggy-backing on the existing
rewrite_aliases pass) when referenced in the rest of the query, to a
unique name based on the name of the query overall to avoid conflicts
w ith other queries that might use the same name for a CTE of their own.

This was... surprisingly easy?