Support for randomly generated data in logictests.
This adds a new `random` command-line flag to `generate`
logictests to randomly generate column data within the
size of the datatype. For example, a TEXT column value
now generates up to 64kB of data. This lets us verify the
correctness of the system with more realistic data structure.

We currently do not generate data larger than 64kB to keep the
in-memory footprint quite small.