Add a TableOperation to delete an entire row
When replicating data (eg from the mysql binlog) we receive updates to
rows in the form of the entire previous row, followed by the entire new
row. We want to turn this into a pair of Negative and Positive records,
but previously base tables only supported issuing direct updates or
deletes based on a primary key. This commit adds a new TableOperation
called DeleteRow, which takes an *entire* row and issues *one* deletion
for that row (eg, if there are multiple instances of that row we will
only delete one of them).