Add new replay_paths_for API, mostly replacing provenance_of
The old provenance_of function in keys.rs was unable to use the new
information generated by the ColumnSource API (and was also somewhat
hard to understand). Additionally, it operated in terms of individual
columns and not indices.

We introduce a new API, replay_paths_for, which makes use of the new
APIs. Salient features:

- instead of being based around (NodeIndex, Vec<Option<usize>>), the new
  API uses the `OptColumnRef` type (NodeIndex, Option<Vec1<usize>>),
  meaning callers don't have to awkwardly unwrap the Options
- the new API will generate "split replay paths" for generated columns;
  replay paths will be generated up to the node with generated columns,
  and then a new set will be started at that node to provide the paths
  needed to implement the handle_upquery API
- this functionality is only used for partial materialization, and can
  be disabled for full materialization
- the new API is split into multiple functions (deduce_column_source /
  continue_replay_path), to make it somewhat easier to read through
- the functions actually have doc comments explaining what they're doing
  this time round :P

The tests in keys.rs are updated to use the new API.