Enable filtering for reader nodes with partial materialization.
During the final stages of setting up materializations, for any reader node that
 is partially materialized, we traverse the dataflow graph upwards until we find
  the first egress node and notify it that it should enable packet filter for
  the reader's domain.
Since a reader nodes belong to its own domain (which means its domain is
only composed of the reader node itself and an ingress node), we only need to
look at the reader node parent (to find the ingress node) and to the ingress
node parent (to find the egress node).
Then, we send a message to the egress node's domain to enable packet filtering
for any packet meant for the ingress node.

However, it might be possible that the ingress node parent is not an egress
node, but instead it is a shard merger.
In that case, we skip setting up the packet filter.
This will be addressed in https://readysettech.atlassian.net/browse/ENG-183