Full-database snapshots for replication
Implement full database replication for mysql connector.
This adds the `--snaphot` option to the to the mysql connector,
that makes it replicate the entire database to Noria and after
that is done, keep reading from the last position in the binlog
file.

The replication follows more or less the same methods used by
debezium and mysqldump.

First a read lock is acquired, then the binlog position is read,
then the recipe for the database is recreated and installed in
Noria, followed by the replication of each individual table.

Fixes #[ENG-158]

https://readysettech.atlassian.net/browse/ENG-158