Store the log name alongside replication offsets
Rather than just storing a single usize for replication offsets, also
store a "replication_log_name" field (as part of a new ReplicationOffset
struct), which if different makes replication offsets not comparable.
In the case of at least MySQL, this will be used to store the *prefix*
of the binlog that we are replicating from, so that if that prefix ever
changes (due to a user changing the mysql option) we don't continue to
try to replicate from the (now intractable) log. There's a new error
type that is already used if replication offsets with different log
names are given as part of the same batch of TableOperations, but which
will *also* be used at *read time* to represent the error described
above.