Start of a multi-process cluster test framework.
This change introduces `clustertest` a multi-process test
framework for managing local test clusters. This allows
creating test clusters for tests and providinhg clients for
issuing requests and collecting metrics. This will allow us to
more comprehensivly test ReadySet deployment behavior while
still remaining small enough to run in CI.

This is the first of several changes to setup the framework.
This commit encompasses defining a very minor API for
clustertest users and creating the zookeeper resources for the
cluster. Future changes will build and start noria-server
binaries according to deployment parameters, and create
metrics clients to check and aggregate cluster metrics in tests.