Fix setting replication offsets on pkey tables
The part of the Base::process method for base tables that have primary
keys was originally written with the assumption that it would only be
passed table operations with valid keys - this assumption was broken in
the case that the set of table operations passed consisted *only* of
keyless operations, such as SetReplicationOffset - eventually we'd try
to call KeyType::from on the key of one of these operations, and that'd
panic. This adds an extra block before the normal handling of
TableOperations to specially handle the case where the set of table
operations to process consists only of keyless operations.

Fixes ENG-184