Update DataType::Real to include precision amount.
This commit updates DataType::Real to include a third enum field of the
precision amount that was recorded at time of creation. This allows us
to better model MySQL behavior, in which MySQL will not give greater
precision back to a user, than the implicit precision that was put into
the system. This also updates all consumers accordingly, and ensures
that we pass this precision in for appropriate string formatting.