Added logic to skip leader election for non-primary region Workers.
During campaign election, if a Worker's region does not match the primary
region, then the Worker will skip the leader election, will wait for a given
backoff period (currently hardcoded at 5 seconds) and go back to checking if
there's a new leader elected.
Also, in order to eventually give up, a retry mechanism is in place (which right
 now is hardcoded to 5 retries, but we can change this down the line to make it
 configurable).
If, after running out of retries, no Controller is found, we assume that no
Noria server is running in the primary region, and exit with an error.