[ENG-162] Clustertest support for starting up noria-server processes.
This commit adds deployment of noria-server instances in the
multi-process cluster. This makes this deployment a full-fledged
noria deployment that can respond to controller requests, queries,
and report metrics. A controller handle and a metrics client have
been added to the deployment handle.

A minimal example test has been added to `clustertest/lib.rs`.
Followup changes will include (1) setting up a test file to add
clustertests, (2) ergonomics changes to make the API easier to
interact with.