Persist replication offsets for schema
Currently replication offset are stored for each table tracking
the next binlog position at the moment the table was updated.
However binlog also keeps track of schema change events that are
not neccesarily attached to any table, such as `CREATE VIEW`.
To prevent such events from being applied more than once, we need
to persist the replication offset for those events with the recipe.

The global offset will therefore be the maximum of the recipe and
the tables.

In addition properly encode the binlog suffix length in `ReplicationOffset`
to enable round trip conversion.