Support for server management (startup/removal) in clustertest.
This commit adds a simple API to startup and kill servers during
a clustertest deployment. A handle to each server can be retrieved
via it's external address.

This change also adds a new controller rpc endpoint
/healthy_workers which returns the url's of the healthy workers
in the deployment. This is used to verify that workers have been
killed as killed workers may not always be immediately removed
from the controller's worker set.

Packged in this commit are several QOL fixes:
  - Using Url's instead of Strings.
  - A cleaner way to build ServerParams in tests.