Retry retrieving STATE_KEY during leader election.
Leader election with a zookeeper authority can enter a state
where CONTROLLER_KEY exists but STATE_KEY does not when starting
up a new deployment. Leader election in a new system occurs when
a noria-server calls `authority.become_leader()`, writing to
CONTROLLER_KEY, and then a write to the STATE_KEY. If a new
noria-server goes through leader election at the same time
it's possible to see that there is a leader in the system but
attempt to retrieve STATE_KEY and see that there is no entry
for the key.

The existing code assumes this is not possible and unwraps the
STATE_KEY value. This commit adds retries the process of reading
STATE_KEY up to 5 times if the node does not exist.