Metric for counting the number of view queries processed and results.
This change records a new metric SERVER_VIEW_QUERY_RESULT, which
is a counter that is incremented with each view query. The tag
`result` is an enum that identifies whether or not the view query
was served directly from cache without upquery or if an upquery
was required.

This also makes two quality-of-life change for testing metrics:
(1) Adding a macro to get a specific metric with specific labels
from a metrics dump, and (2) adding querying specific workers for
their metrics dump to the MetricsClient.