ast: Merge ConditionExpression into Expression
This commit, which represents (probably, for now) the final piece in the
series of refactors and simplifications to the nom-sql AST refactor that
included 70f35e60c, c1c666f21, d6f514e1d, and cf665d265, merges
ConditionExpression entirely into Expression, along the way removing
the (completely artificial) distinction between LogicalOp, ComparisonOp,
and Arithmetic, instead electing to just have a single `BinaryOp` AST
node that represents all binary operator applications. This is done
because it's generally considered best-practice for an AST to represent
only the *syntax* of the code, and not any of the *semantics*, and leave
determining semantics (such as differentiating between comparisons and
logical operations, for example) to later phases.

In order to properly parse operator precedence, this continues with the
usage of the `pratt` crate that was introduced in the refactor of the
Arithmetic type in 70f35e60c, extending it to support both all binary
operators and the one unary operator we currently support, NOT.

Some more highlights of what's going on to make all this work in
this (admittedly very, very large) commit:

- Since ConditionExpression has gone away entirely, all of the previous
  tests in nom_sql::condition (which is also gone) have moved to
  nom_sql::expression
- Previously, we'd convert the IS operator to an `=` operator *during
  parsing* (yikes!) - this is especially bad since they actually
  have *different semantics* with regards to NULL - I've opened a
  ticket (ENG-209) to fix our handling of NULL comparisons, and moved
  the broken behavior that treats those two operators identically to
  eval-time
- The structure of the IN operation changed slighty upon perusal of the
  ANSI SQL standard: it's actually only syntacially valid for an IN to
  have either a subquery or a comma-separated parenthesized list of
  expressions as its right-hand side! Given this, IN is actually its own
  enum variant in Expression now instead of a binary operator, with its
  own enum to handle those two things that can appear on its right-hand
  side - and ConditionExpression::LiteralList has gone away entirely to
  make that work. To support this, there have been a few changes to the
  subquery extraction pass to explicitly differentiate between a
  subquery that appears as an Expression, and a subquery that appears on
  the right-hand side of an IN.
- A couple of things were dealing with raw `ConditionTree`s - there's no
  equivalent type for that here, since `Expression::BinaryOp` is a
  top-level struct enum variant - most of the time that's ok and we can
  just pass around an `Expression` throwing an `unsupported!()` error in
  the case that it's not a `BinaryOp` (this is good, even, as it paves
  the way for us doing predicates on things that aren't just a binary
  operator, such as function calls that return a boolean) - the one
  place I've elected to *not* do that is join predicates, as the
  assumption that a join predicate is a binary equality comparison
  between columns on each table is pretty embedded into eg the query
  graph, and lifting that limitation is going to require singificantly
  more design. To that end, I've introduced a new JoinPredicate type in
  query_graph itself, which is only used there.
- Since we're now walking more parts of expressions that might refer to
  the same table twice, the return type of `referred_tables` in
  `nom_sql::analysis` has changed from a Vec to a HashSet

One final thing that this commit *doesn't* do, but makes considerably
easier than it used to be, is replace all of the ad-hoc traversals of
the Expression AST that happen during analysis and rewrite passes with a
visitor[0] - since we now only have one expression AST, we can write
only one Visitor trait for that AST, and implement the logic for pattern
matching and recursively traversing the AST only once, decoupling it
from the actual logic of the individual analysis and rewrite passes
themselves.

[0]: https://rust-unofficial.github.io/patterns/patterns/behavioural/visitor

Fixes: ENG-52